# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Doc Index API (Local)",
        "version": "0.1.0"
    },
    "paths": {
...
        "/search": {
            "get": {
                "tags": [
                    "Busca"
                ],
                "summary": "Search",
                "operationId": "search_search_get",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "description": "Consulta FTS5 (use aspas para frase, AND/OR, NEAR)",
                            "title": "Q"
                        },
                        "description": "Consulta FTS5 (use aspas para frase, AND/OR, NEAR)"
                    },
                    {
                        "name": "ext",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Filtro por extensões: ex 'pdf,docx,txt'",
                            "title": "Ext"
                        },
                        "description": "Filtro por extensões: ex 'pdf,docx,txt'"
                    },
                    {
                        "name": "root_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "integer"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Filtrar por uma raiz específica",
                            "title": "Root Id"
                        },
                        "description": "Filtrar por uma raiz específica"
                    },
                    {
                        "name": "since",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "YYYY-MM-DD",
                            "title": "Since"
                        },
                        "description": "YYYY-MM-DD"
                    },
                    {
                        "name": "until",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "YYYY-MM-DD",
                            "title": "Until"
                        },
                        "description": "YYYY-MM-DD"
                    },
                    {
                        "name": "min_size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "integer",
                                    "minimum": 0
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Min Size"
                        }
                    },
                    {
                        "name": "max_size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "integer",
                                    "minimum": 0
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Max Size"
                        }
                    },
                    {
                        "name": "project",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Código do projeto para boost",
                            "title": "Project"
                        },
                        "description": "Código do projeto para boost"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 500,
                            "minimum": 1,
                            "default": 50,
                            "title": "Limit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/SearchResult"
                                    },
                                    "title": "Response Search Search Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/download": {
            "get": {
                "tags": [
                    "Download"
                ],
                "summary": "Download",
                "operationId": "download_download_get",
                "parameters": [
                    {
                        "name": "path",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "description": "Caminho absoluto (UNC ou local)",
                            "title": "Path"
                        },
                        "description": "Caminho absoluto (UNC ou local)"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/files": {
            "get": {
                "tags": [
                    "Arquivos (Metadados)"
                ],
                "summary": "List Files",
                "operationId": "list_files_files_get",
                "parameters": [
                    {
                        "name": "root_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "integer"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Root Id"
                        }
                    },
                    {
                        "name": "ext",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "ex.: pdf,docx (uma extensão)",
                            "title": "Ext"
                        },
                        "description": "ex.: pdf,docx (uma extensão)"
                    },
                    {
                        "name": "min_size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "integer",
                                    "minimum": 0
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Min Size"
                        }
                    },
                    {
                        "name": "max_size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "integer",
                                    "minimum": 0
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Max Size"
                        }
                    },
                    {
                        "name": "min_mtime",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "integer",
                                    "minimum": 0
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "epoch seconds",
                            "title": "Min Mtime"
                        },
                        "description": "epoch seconds"
                    },
                    {
                        "name": "max_mtime",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "integer",
                                    "minimum": 0
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "epoch seconds",
                            "title": "Max Mtime"
                        },
                        "description": "epoch seconds"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 1000,
                            "minimum": 1,
                            "default": 100,
                            "title": "Limit"
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 0,
                            "default": 0,
                            "title": "Offset"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/FileOut"
                                    },
                                    "title": "Response List Files Files Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/files/{file_id}": {
            "get": {
                "tags": [
                    "Arquivos (Metadados)"
                ],
                "summary": "Get File",
                "operationId": "get_file_files__file_id__get",
                "parameters": [
                    {
                        "name": "file_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "File Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FileOut"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "FileOut": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "root_id": {
                        "type": "integer",
                        "title": "Root Id"
                    },
                    "path": {
                        "type": "string",
                        "title": "Path"
                    },
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "ext": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Ext"
                    },
                    "size": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Size"
                    },
                    "mtime": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Mtime"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "root_id",
                    "path",
                    "name",
                    "ext",
                    "size",
                    "mtime"
                ],
                "title": "FileOut"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "SearchResult": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "ext": {
                        "type": "string",
                        "title": "Ext"
                    },
                    "path": {
                        "type": "string",
                        "title": "Path"
                    },
                    "file_uri": {
                        "type": "string",
                        "title": "File Uri"
                    },
                    "download_url": {
                        "type": "string",
                        "title": "Download Url"
                    },
                    "size": {
                        "type": "integer",
                        "title": "Size"
                    },
                    "mtime": {
                        "type": "string",
                        "title": "Mtime"
                    },
                    "score": {
                        "type": "number",
                        "title": "Score"
                    },
                    "snippet": {
                        "type": "string",
                        "title": "Snippet"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "ext",
                    "path",
                    "file_uri",
                    "download_url",
                    "size",
                    "mtime",
                    "score",
                    "snippet"
                ],
                "title": "SearchResult"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}